==> 03_Imputing Drug Response in GDSC.Rmd <==
---
title: "Build models using the CCLE/CTRPv2 datasets to impute drug response in GDSC cell lines"
author: "Robert F. Gruener"
output:
  html_document:
    df_print: paged
date: "date last edited: `r format(Sys.time(), '%d %B, %Y')`"
---
```{r, include=FALSE}
knitr::opts_chunk$set(warning = F, message = FALSE, cache = T)
```
### Objectives

As a cross-database-validation, the goals are to compare imputed results from CTRP with measured results from GDSC. To do this: 
 * Set the parameters
 * Load the GDSC test expression data
 * Load the CTRP/CCLE Training expression and drug data
 * Generate the models and impute using the pRRophetic framework

Important: Set the parameters appropriately, then everything else *should* run smoothly if you run the enitre file. 

### First, load packages and set parameters
```{r loading packages and setting parameters, results= "hide"}
set.seed(12345)
#Loading Packages
needed_packages <- c("tidyverse", "readxl", "glmnet", "gdata", "illuminaHumanv4.db" ,"car", "ridge", "preprocessCore", "genefilter", "sva", "pRRophetic")
lapply(needed_packages, require, character.only = TRUE)

###################### Set Parameters ##########################

# Set location and filename  for list/matrix R.data output 
dir.create("DataOut")
#list_output_name <- "./DataOut/CTRPbreastcancerPREDS.RData" #we won't be using the list, but you can save it if you so choose 
matrix_output_name <- "./DataOut/CTRP-models_in_GDSC.Rdata"

# Finally, set parameters of pRRophetic's calcphenotype function 
#batchCorrect options: "eb" for ComBat, "qn" for quantiles normalization, "standardize", or "none"
batchCorrect <- "standardize"
#powerTransformPhenotype: does a power transformation to the phenotype data, recommended over log transformations by the literture for use in Ridge regression 
powerTransformPhenotype <- TRUE
#removeLowVaryingGenes: percentage of low varying genes to remove, default is at 0.2 
removeLowVaryingGenes <- 0.2
#removeLowVaryingGenesFrom: where should it remove low varying genes from "homogenizeData" or "rawData"
removeLowVaringGenesFrom <- "rawData"
#minNumSamples: minimum number of training samples, set to 30, but this shouldn't be an issue since the there are many samples 
minNumSamples = 30
#selection: how to deal with duplicated gene IDs, -1 for ask user, 1 for summarize by mean, 2 for disregard duplicates
selection <- 1
#printoutput: should it print outputs, default is TRUE
printOutput = TRUE

convert.to.matrix <- function(df){
  mat <- as.matrix(df[,-1])
  rownames(mat) <- c(t(df[,1]))
  return(mat)
  }

```

### Second, GDSC expression data

```{r Load RNA-Seq}
############ Load Test Expression data (as Matrix) ###################

GDSC_2016_Expression_Data <- read.delim("./DataIn/GDSC/GDSC_2016expression_clean.txt", sep = "\t", stringsAsFactors = F)    

testExprData <- GDSC_2016_Expression_Data
colnames(testExprData) <- paste(colnames(GDSC_2016_Expression_Data), "test", sep = "_") 
#The Test expression dataset cannot have the same names as the traning expression dataset. However, cell line names (colnames) will overlap between GDSC and CTRP, so adding "test" to the column names prevents this issue. The imputation framework below only does batch correction across all the cell lines; luckily, the models are built for every test subject separately and independently. That is, we can still perform an analysis of non-overlapping cell lines even if we impute in all. Additionally, the expression data is coming from two completely independent datasets. For these 2 reasons, we will impute in all the cell lines even if they overlap with CTRP. 

testExprData <- convert.to.matrix(testExprData)

```

### Third, load CTRP training data

Create models using CTRP for imputations in the GDSC RNA-Seq as test expression

```{r CTRP Imputations on GDSC RNA-seq data, results = "hide"}

#Loads training phenotype data 
CTRPv2_AUC <- read.delim("./DataIn/CTRPv2/CTRPv2_AUC_clean.txt", sep = "\t", stringsAsFactors = F)
#Altering cell line names that start with a number to start with an X to match rownames of expression data
start.with.number <- grep("^[[:digit:]]", CTRPv2_AUC$CCL_Name)
CTRPv2_AUC$CCL_Name[start.with.number] <- paste("X", CTRPv2_AUC$CCL_Name[start.with.number], sep = "")
trainingPhenData <- CTRPv2_AUC

#Load Training Expression data, convert it to a matrix
CTRPv2_RNAseq_TPM <- read.delim("./DataIn/CTRPv2/CTRPv2_RNAseq_TPM_clean.txt",  sep = "\t", stringsAsFactors = F)

ExprData <- convert.to.matrix(CTRPv2_RNAseq_TPM)
```


### Fourth, determine drug overlap

There is no point in performing the imputations on all the CTRP drugs since we are only interested in the overlap between GDSC and CTRP. 

```{r Find overlap between CTRP and GDSC}

GDSC_drugs <- read_xlsx("./DataIn/GDSC/GDSC_IC50_July2019.xlsx")

GDSC_drugs_vec <- unique(GDSC_drugs$DRUG_NAME)
harmonizing_dataset <- read_csv("./Included_Data/Drug_targets(review).csv")

harm_GDSC_drug_vec <- unique(harmonizing_dataset$`Harmonized Name`[harmonizing_dataset$`Name in Database` %in% GDSC_drugs_vec])

CTRP_names_overlap_with_GDSC <- harmonizing_dataset %>% 
  filter(Database == "CTRPv2") %>% 
  filter(`Harmonized Name` %in% harm_GDSC_drug_vec) %>% 
  .$`Name in Database`

possibleDrugs <- CTRP_names_overlap_with_GDSC 

```


### Fifth, perform imputations


```{r Do the imputations, cache= TRUE, autodep= TRUE}
#make a list of all possible drugs predictions

keep(ExprData, possibleDrugs, trainingPhenData, testExprData, batchCorrect, powerTransformPhenotype, 
     removeLowVaringGenesFrom, minNumSamples, selection, printOutput, removeLowVaryingGenes, matrix_output_name, sure = TRUE)
CTRPDrugPredictions <- list()


########make a for loop for each drug seperate#################
for(i in 1:length(possibleDrugs)) { 
  drug <- possibleDrugs[i]
  temp <- trainingPhenData[trainingPhenData$cpd_name == drug,] ## Makes temp matrix to extract the AUCs easier
  AUCs <- as.numeric(temp$Avg_AUC)
  names(AUCs) <- temp$CCL_Name
  commonCellLines <- colnames(ExprData)[colnames(ExprData) %in% names(AUCs)]
  AUCsOrd <- AUCs[commonCellLines]
  trainDataOrd <- ExprData[, commonCellLines]
  trainDataOrd <- trainDataOrd[(rowSums(trainDataOrd) != 0), ]

  CTRPDrugPredictions[[i]] <- calcPhenotype(trainingExprData =  trainDataOrd, 
                                trainingPtype = AUCsOrd, testExprData = testExprData, batchCorrect = batchCorrect, 
                                powerTransformPhenotype = powerTransformPhenotype, 
                                removeLowVaryingGenes = removeLowVaryingGenes, 
                                minNumSamples = minNumSamples, selection = selection, 
                                printOutput = printOutput, removeLowVaringGenesFrom = removeLowVaringGenesFrom)
} #close for loop

names(CTRPDrugPredictions) <- possibleDrugs
CTRPDrugPredictions_mat <- do.call(rbind, CTRPDrugPredictions)
# save(CTRPDrugPredictions, file= list_output_name)
save(CTRPDrugPredictions_mat, file= matrix_output_name)



```

```{r}
print(sessionInfo())
```
