==> 03A_Imputing_Cross Validation of CTRP.Rmd <==
---
title: '20-Fold Cross-validation on CTRP-models'
output: html_notebook 
author:
  - "Robert F. Gruener"
date: "date last edited: `r format(Sys.time(), '%d %B, %Y')`"
---

### Objectives

Do the imputations

 * Set the parameters
 * Load the test expression data (TCGA Breast Cancer)
 * Load the CTRP/CCLE Training expression and drug data
 * Run the cross-validation 



```{r, include=FALSE}
knitr::opts_chunk$set(message = F, warning = F, cache = T)
```

### First, load packages and set parameters

The necessary packages and files should have been download in file 01_Downlaod

```{r loading packages and setting parameters, results= "hide"}
set.seed(12345)
#Loading Packages
needed_packages <- c("tidyverse", "readxl", "glmnet", "gdata", "illuminaHumanv4.db" ,"car", "ridge", "preprocessCore", "genefilter", "sva", "pRRophetic")
lapply(needed_packages, require, character.only = TRUE)

# library(doParallel), could easily be made to run in parallel

```

### Second, load the RNA-Seq data

```{r Load RNA-Seq}
############ Load Test Expression data (as Matrix) ###################
# load  the breast cancer RNA-seq file downloaded previously
tpmDatMat_tpm_logged <- read.delim(file = "./DataIn/TCGA/TCGA_Breast_TPM.txt", as.is=T)

```

### Third, load and clean the CTRP/CCLE data:

Perform imputations using CTRP/CCLE datasets as the training dataset

```{r CTRP Imputations, message = FALSE}

#Loads training phenotype data 
CTRPv2_AUC <- read.delim("./DataIn/CTRPv2/CTRPv2_AUC_clean.txt", sep = "\t", stringsAsFactors = F)
#Altering cell line names that start with a number to start with an X to match rownames of expression data
start.with.number <- grep("^[[:digit:]]", CTRPv2_AUC$CCL_Name)
CTRPv2_AUC$CCL_Name[start.with.number] <- paste("X", CTRPv2_AUC$CCL_Name[start.with.number], sep = "")
trainingPhenData <- CTRPv2_AUC
#names(trainingPhenData)[names(trainingPhenData)== "cpd_name"] <- "Drug.name"
possibleDrugs <- unique(trainingPhenData$cpd_name)

#Load Training Expression data, convert it to a matrix
CTRPv2_RNAseq_TPM <- read.delim("./DataIn/CTRPv2/CTRPv2_RNAseq_TPM_clean.txt",  sep = "\t", stringsAsFactors = F)
convert.to.matrix <- function(df){
  mat <- as.matrix(df[,-1])
  rownames(mat) <- c(t(df[,1]))
  return(mat)
  }
ExprData <- convert.to.matrix(CTRPv2_RNAseq_TPM)



```

### Fourth, do cross-validation


Note for the cross-validation. It takes up a lot of memory, especially if you don't subset the genes. There are 17,000 genes that overlap between training and the test data, but if you j ust use the training data there are 50,000+ genes which usually (on a base computer) will run into a protect() stack overflow error. For running the LOOCV with all the cell lines (for MK-1775, all 762 of them), you will likely run into an error of the vector being too large (6 GB) to store. 


### A more efficient CV

```{r}

#free up some memory 
rm(list = keep(trainingPhenData, ExprData, testExprData, mk1775))

possibleDrugs <- unique(trainingPhenData$cpd_name)

corP_spear <- numeric()
cor_spear <- numeric()
corP_pearson <- numeric()
cor_pearson <- numeric()
RMSEs <- numeric()
nAll <- numeric()

rmse <- function(error){
  sqrt(mean(error^2, na.rm = T))
}


########make a for loop for each drug seperate#################
for(i in 1:length(possibleDrugs)) { 
  drug <- possibleDrugs[i]
  temp <- trainingPhenData[trainingPhenData$cpd_name == drug,] ## Makes temp matrix to extract the AUCs easier
  AUCs <- as.numeric(temp$Avg_AUC)
  names(AUCs) <- temp$CCL_Name
  AUCs <- AUCs
  commonCellLines <- colnames(ExprData)[colnames(ExprData) %in% names(AUCs)]
  AUCsOrd <- AUCs[commonCellLines]
  trainDataOrd <- ExprData[, commonCellLines]
  trainDataOrd <- trainDataOrd[(rowSums(trainDataOrd) != 0), ]
  CTRP_LOOCV <- predictionAccuracyByCv(trainingExprData =  trainDataOrd, trainingPtype = AUCsOrd, testExprData = testExprData, cvFold = 20)
  # correlations on all tumors.
    spearCor <- cor.test(CTRP_LOOCV[[1]], CTRP_LOOCV[[2]], method = "spearman")
    corP_spear[i] <- spearCor$p.value
    cor_spear[i] <- spearCor$estimate
    
    nAll[i] <- length(CTRP_LOOCV[[1]])
    
    pearson_Cor <- cor.test(CTRP_LOOCV[[1]], CTRP_LOOCV[[2]], method = "pearson")
    corP_pearson[i] <- pearson_Cor$p.value
    cor_pearson[i] <- pearson_Cor$estimate
    error <- CTRP_LOOCV[[1]] - CTRP_LOOCV[[2]]
    RMSEs[i] <- rmse(error)

  
} #close for loop

m <- cbind(possibleDrugs, nAll, RMSEs, cor_spear, corP_spear, cor_pearson, corP_pearson)
colnames(m) <- c("Drug", "N_CCLs","RMSE", "Spearman_Correlation" , "Spearman_Pvalue", "Pearson_Correlation", "Pearson_Pvalue")

write.csv(m, file = "./DataOut/20FoldCV_CTRP.csv")
```


```{r}
print(sessionInfo())
```
