==> 04_Analyze_Cell-Line-Only-Analysis.Rmd <==
---
title: "04_Analyze_Cell-Line-Only-Analysis"
output: html_notebook
author:
  - "Robert F. Gruener"
date: "date last edited: `r format(Sys.time(), '%d %B, %Y')`"
---
```{r, include=FALSE}
knitr::opts_chunk$set(warning = F, message = FALSE, cache = T)
```


### Cell Line Only Analysis

The breast cancer cell line info is from [a recent review](http://www.jcancer.org/v08p3131.htm) and shows the molecular classifications of the breast cancer cell lines. This file is also included in the repo for ease of access. 

Shown below is some of the information these data files contain. 

```{r}
library(tidyverse)
library(readxl)
library(broom)

theme_set(theme_bw())
```


```{r data_loading, cache=TRUE}
print("Breast cancer cell line info")
breast_cell_line_info <- read_csv(file = "./Included_Data/Breast_CCL_Subtypes.csv")
cell_lines <- read_excel("./Included_Data/All Cell Line Info.xlsx")
ccls <- cell_lines %>% select(`Harmonized Identifier`, `Study Cell Line Name`, `Cancer Statistics 2017 Sub-Category`)
```

##### Cleaning and Subsetting the data

We now need to combine the info files with the drug data, and subset to just the breast cancer cell lines

```{r clean_subset}

GDSC_IC50s <- read_excel("./DataIn/GDSC/GDSC_IC50_July2019.xlsx")
GDSC_breast <- GDSC_IC50s %>% 
  select(CELL_LINE_NAME, DRUG_NAME, LN_IC50) %>%  #selecting relevant columns 
  left_join(., ccls, by = c("CELL_LINE_NAME" = "Study Cell Line Name")) %>% 
  distinct() %>% 
  filter(`Cancer Statistics 2017 Sub-Category` == "Breast") %>% #filter to breast cancers only
  #need to remove - from cell line names to make matches easier, chose to remove all punctuation as follows:
  mutate(CCL_Name = gsub("[[:punct:]]", replacement = "", x = CELL_LINE_NAME)) %>% 
  #need to join with breast cancer subtype data
  left_join(., breast_cell_line_info, by = c("CCL_Name" = "Cell lines"))

CTRP_aucs <- read_delim("./DataIn/CTRPv2/CTRPv2_AUC_clean.txt", delim = "\t")
CTRP_breast <- CTRP_aucs %>%
   left_join(., ccls, by = c("CCL_Name" = "Study Cell Line Name")) %>%
   distinct() %>%
   drop_na(`Harmonized Identifier`) %>%
   filter(`Cancer Statistics 2017 Sub-Category` == "Breast") %>%
   select(CCL_Name, cpd_name, apparent_ec50_umol, Avg_AUC, `Harmonized Identifier`) %>%  #selecting relevant columns
   #need to join with breast cancer subtype data
   left_join(., breast_cell_line_info, by = c("CCL_Name" = "Cell lines"))

```

Unfortunately, there is likely some missing subtype information still. Let's see:

``` {r clean_subset_missing}
# Unfortunately, there is likely some missing subtype information still. Let's see:

missing_gdsc <- GDSC_breast %>% 
    select(CCL_Name, Subtype) %>% 
    filter(is.na(Subtype)) %>% 
    unique() 

CTRP_breast %>% 
    select(CCL_Name, Subtype) %>% 
    filter(is.na(Subtype)) %>% 
    unique() %>% 
    bind_rows(., missing_gdsc) %>% 
    unique()
    

```

Those 10 values will be manually curated. 

```{r clean_subset_2}
#I manually curated 12 values previously, so I'll use these 
# these sites helped with the manual curation, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3127435/, https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5001206/

missing_subtype <- c("TNB", "H", "LA", "UN", "TNB", "UN", "TN", "LB", "TN", "H", "UN", "TNB") #"UN", unclassifiable, problematic cell lines with conflicting status claims in the literature
names(missing_subtype) <- c("HDQP1", "JIMT1", "HMC18", "HCC1500", "HS578T", "HCC1419", "COLO824", "EVSAT", "MRKnu1", "OCUBM", "YMB1E", "MB157")

GDSC_breast <- GDSC_breast %>% 
  mutate(Subtype= if_else(is.na(Subtype), missing_subtype[CCL_Name], Subtype)) %>% 
  mutate(TNBC = ifelse(Subtype %in% c("TN", "TNA", "TNB"), "TNBC", "RPBC"))

CTRP_breast <- CTRP_breast %>% 
  mutate(Subtype= if_else(is.na(Subtype), missing_subtype[CCL_Name], Subtype)) %>% 
  mutate(TNBC = ifelse(Subtype %in% c("TN", "TNA", "TNB"), "TNBC", "RPBC"))

```


With that, now we will get into the statistics. Below are the significant results for both datasets:

```{r compare_results_statistics}
GDSC_t.test_results <- GDSC_breast %>% 
  group_by(DRUG_NAME) %>% 
  do(tidy(t.test(LN_IC50 ~ TNBC, data = .))) %>% # thanks second answer here
  # https://stackoverflow.com/questions/26244321/dplyr-summarise-multiple-columns-using-t-test
  ##I now need to change names because it is unclear what the values are. 
  #it goes in alphabetical order, so Estimate is the mean IC50 when TNBC=="no" minus the mean of TNBC=="yes" 
  ungroup() %>% 
  mutate(FDR = p.adjust(p.value, method = "fdr"), 
         more_sensitive = if_else(estimate > 0, "TNBC", "non-TNBC")) #this summarizes 
          ##the effect direction. The smaller the number, the more sensitive the cell was to this drug
          ###So if the values are positive it means TNBC was more negative and thus more sensitive

GDSC_t.test_results %>% arrange(FDR)



```

As we can see, there are `r length(GDSC_t.test_results$FDR[GDSC_t.test_results$FDR < 0.05])` significant results in GDSC after multiple comparision correction. 

Also, it is interesting to note that typically the direction is in favor of TNBC being *more sensitive*. In GDSC, even if we  use a p.value of 0.05 (note, FDR not p-value), there are `r sum(GDSC_t.test_results[GDSC_t.test_results$FDR < 0.05,]$more_sensitive == "TNBC")` drugs that are more effective on TNBC cells and only `r sum(GDSC_t.test_results[GDSC_t.test_results$FDR < 0.05,]$more_sensitive == "non-TNBC")` shown to be more effective on the RPBC cells. 


```{r}
#Let's Plot the Result



GDSC_breast %>% 
  filter(DRUG_NAME == "MK-1775") %>% 
  ggplot(aes(x = TNBC, y = LN_IC50)) + 
    geom_boxplot() + 
    geom_jitter(width = 0.2, aes(color = TNBC)) +
    labs(title = "GDSC Breast CCL Measured AZD-1775 Response Stratified by TNBC Status", 
         y = "Measured IC50 to AZD-1775", 
         x = NULL) + 
  theme(legend.position = "none") + 
  geom_hline(yintercept = log(1.65))



#There is one breast cancer cell line MDA-MB-175-VII or CVCL_1400 that wasn't imputed on apparently. Perhaps that one is missing RNA-Seq Data
 

```



```{r}
load("./DataOut/MK1775CTRP_in_GDSC.Rdata") #use the matrix_output_name from IMputing Drug Response in GDSC

preds_fromCTRP <- as_tibble(t(CTRPDrugPredictions_mat), rownames = "CCLs") %>% 
  separate(CCLs, sep = "_", into = c("CCL", NA))

predsToMerge <- preds_fromCTRP %>% 
  left_join(., ccls, by = c("CCL" = "Study Cell Line Name")) %>% 
  distinct() %>% #distinct is needed because we are using the Study Cell Line Name which includes multiple studies, so the same name might be listed twice
  drop_na(`Harmonized Identifier`) #there are 8 cell lines that did not match with a Cellosaurus ID, which I am simply dropping for convenience


GDSC_breast %>% 
  filter(DRUG_NAME == "MK-1775") %>% 
  left_join(., predsToMerge, by = "Harmonized Identifier") %>% 
  ggplot(aes(x = TNBC, y = `MK-1775`)) + 
    geom_boxplot() + 
    geom_jitter(width = 0.2, aes(color = TNBC)) +
    labs(title = "GDSC Breast CCL Imputed AZD-1775 Response Stratified by TNBC Status" ,
         y = "AZD-1775 Imputed Sensitivity Score",
         x = NULL) + 
    theme(legend.position = "none")
```


```{r}
CTRP_breast <- CTRP_aucs %>%
  filter(cpd_name %in% c("MK-1775", "olaparib")) %>% 
  left_join(., ccls, by = c("CCL_Name" = "Study Cell Line Name")) %>%
  distinct() %>% 
  drop_na(`Harmonized Identifier`) %>% 
  filter(`Cancer Statistics 2017 Sub-Category` == "Breast", 
         apparent_ec50_umol < 135) %>%  
  select(CCL_Name, cpd_name, apparent_ec50_umol, Avg_AUC, `Harmonized Identifier`) %>%  #selecting relevant columns 
  #need to join with breast cancer subtype data
  left_join(., breast_cell_line_info, by = c("CCL_Name" = "Cell lines"))

missing_subtype <- c("TNB", "H", "LA", "UN", "TNB", "UN", "TN", "LB", "TN", "H", "UN", "TNB") #"UN", unclassifiable, problematic cell lines with conflicting status claims in the literature
names(missing_subtype) <- c("HDQP1", "JIMT1", "HMC18", "HCC1500", "HS578T", "HCC1419", "COLO824", "EVSAT", "MRKnu1", "OCUBM", "YMB1E", "MB157")

CTRP_breast <- CTRP_breast %>% 
  mutate(Subtype= if_else(is.na(Subtype), missing_subtype[CCL_Name], Subtype)) %>% 
  mutate(TNBC = if_else(Subtype %in% c("TN", "TNA", "TNB"), "TNBC", "RPBC"))


CTRP_breast %>% filter(cpd_name == "MK-1775") %>%
mutate(rank = dense_rank(apparent_ec50_umol)) %>% 
  ggplot(aes(x = rank, y = apparent_ec50_umol, color = TNBC)) + 
    geom_point() + 
    geom_hline(yintercept = 1.3, color = "red") + 
    labs(title = "MK-1775")

CTRP_breast %>% filter(cpd_name == "olaparib") %>% 
mutate(rank = dense_rank(apparent_ec50_umol)) %>% 
  ggplot(aes(x = rank, y = apparent_ec50_umol, color = TNBC)) + 
    geom_point() + 
    geom_hline(yintercept = 21, color = "red") + 
    labs(title = "Olaparib")

```






```{r}
GDSC_breast %>%
filter(DRUG_NAME == "MK-1775") %>% 
distinct() %>% 
drop_na(`Harmonized Identifier`) %>% 
mutate(rank = dense_rank(LN_IC50)) %>% 
  ggplot(aes(x = rank, y = LN_IC50)) + 
    geom_point() + 
    geom_hline(yintercept = log(1.6), color = "red")
```


```{r}
print(sessionInfo())
```
