==> 04_Analyze_Imputations-on-GDSC-Data.Rmd <==
---
title: "Analyze: GDSC (Cell Line) Imputations"
output: html_notebook
author:
  - "Robert F. Gruener"
date: "date last edited: `r format(Sys.time(), '%d %B, %Y')`"
---
```{r, include=FALSE}
knitr::opts_chunk$set(warning = F, message = FALSE, cache = T)
```

Analysis outline: 

  * Load AZD1775 imputations in GDSC and the actual measured responses from GDSC
  * Format the data and combine the datasets 
  * Perform correlation analysis between imputed and measured responses 
  * Get cell line subtypes for the GDSC cell lines
  * Subtype the cell lines into TNBC and RPBC CCLs and perform t-test and see if the results also indicate that AZD1775 is more sensitive in TNBC cell lines 


```{r load some stuff, include = FALSE}
library(tidyverse)
library(broom)
library(readxl)
library(magrittr)

load("./DataOut/MK1775CTRP_in_GDSC.Rdata") #use the matrix_output_name from IMputing Drug Response in GDSC

preds_fromCTRP <- as_tibble(t(CTRPDrugPredictions_mat), rownames = "CCLs") %>% 
  separate(CCLs, sep = "_", into = c("CCL", NA))
GDSC_IC50s <- read_excel("./DataIn/GDSC/GDSC_IC50_July2019.xlsx") #read the drug response data downloaded earlier
cell_lines <- read_excel("./Included_Data/All Cell Line Info.xlsx") #read in the Cell Line info from our review that was included in this repo, otherwise needs to be downloaded separately (see Download_TCGA... .Rmd file)
ccls <- cell_lines %>% select(`Harmonized Identifier`, `Study Cell Line Name`, `Cancer Statistics 2017 Sub-Category`)


```


```{r formatting}
#format GDSC IC50 values to prepare for merging with the predictions
GDSCtomerge <- GDSC_IC50s %>% 
  select(CELL_LINE_NAME, DRUG_NAME, LN_IC50) %>% 
  filter(DRUG_NAME == "MK-1775") %>% 
  left_join(., ccls, by = c("CELL_LINE_NAME" = "Study Cell Line Name")) %>% 
  select(-CELL_LINE_NAME) %>% 
  distinct() #distinct is needed because we are using the Study Cell Line Name which includes multiple studies

#format predictions values to prepare for merging with IC50 values
predsToMerge <- preds_fromCTRP %>% 
  left_join(., ccls, by = c("CCL" = "Study Cell Line Name")) %>% 
  distinct() %>% #distinct is needed because we are using the Study Cell Line Name which includes multiple studies, so the same name might be listed twice
  drop_na(`Harmonized Identifier`) #there are 8 cell lines that did not match with a Cellosaurus ID, which I am simply dropping for convenience

#merge two datasets
Merge_preds_GDSC.IC50 <- left_join(predsToMerge, GDSCtomerge, by = "Harmonized Identifier")



```


To begin, after merging the Imputed and measured datasets, make a graph of imputed and measured response to MK-1775 for all of GDSC and only the breast cancer cell lines in GDSC, and see how well they correlate. 

```{r Imputed and Measured Sensitivity Correlation}

theme_set(theme_bw()) #sets the theme for ggplot

#Make graph of imputed and measured response to MK-1775 for all of GDSC
sfig4a <-Merge_preds_GDSC.IC50 %>% 
  drop_na(LN_IC50) %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity in GDSC", 
         x = "Imputed Sensitivty Score",
         y = "Measured ln(IC50) values from GDSC")

sfig4a <- sfig4a + labs(title = NULL)

#Make the same graph of imputed and measured response to MK-1775, but for only the breast cancer cell lines in GDSC
sfig4b <- Merge_preds_GDSC.IC50 %>% 
  drop_na(LN_IC50) %>% 
  filter(`Cancer Statistics 2017 Sub-Category.x` == "Breast") %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Actual AZD-1775 Sensitivity in GDSC Breast CCLs",
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) values from GDSC")

sfig4b <- sfig4b + labs(title = NULL)

```

As we can see, they correlate very strongly. However, even though the CTRP and GDSC datasets are completely independent, some of these cell lines overlap between the datasets and therefore appear in both the training and testing datasets. Thus, it may be important to look at correlation between cell lines that are unique to GDSC, that is those that are truly independent, to make sure that the same trend holds.  

```{r Unique to GDSC Analysis}

CTRP_aucs <- read_delim("./DataIn/CTRPv2/CTRPv2_AUC_clean.txt", delim = "\t") #read in data used to make the imputations and downloaded in the download... .Rmd file

#format
Ccls_CTRP <- CTRP_aucs %>% 
  filter(cpd_name == "MK-1775") %>% 
  left_join(., ccls, by = c("CCL_Name" = "Study Cell Line Name")) %>% 
  distinct() %>% 
  drop_na(`Harmonized Identifier`)

#check to make sure nothing is missing
CTRP_aucs %>% filter(cpd_name == "MK-1775") %$% length(.$cpd_name) == 
  length(na.omit(Ccls_CTRP$`Harmonized Identifier`))

#use anti_join to find the CCLs unique to GDSC, ie removing ccls in GDSC that are also in CTRP
non_modeled_CCLs <- anti_join(GDSCtomerge, Ccls_CTRP, by = "Harmonized Identifier")

#graph measured and Imputed response in all GSDC cell lines that are unique to GDSC. Probably should have made a graphing function, but oh well. 
figure4a <- Merge_preds_GDSC.IC50 %>% 
  semi_join(., non_modeled_CCLs) %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity \n in Cell Lines Unique to GDSC", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) values from GDSC") + 
    theme_bw(base_size = 14.6)
figure4a
figure4a <- figure4a + labs(title = NULL)

#graph measured and Imputed response in only BRCA GSDC cell lines that are unique to GDSC, n = 16

sup_fig4c <- Merge_preds_GDSC.IC50 %>% 
  semi_join(., non_modeled_CCLs) %>% 
  filter(`Cancer Statistics 2017 Sub-Category.x` == "Breast") %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity \n in Breast Cancer Cell Lines Unique to GDSC", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) values from GDSC")
sup_fig4c
sup_fig4c <- sup_fig4c + labs(title = NULL)

```


As we can see, even in the unique to GDSC cell lines, the Imputed and measured response correlates well. 

Now these correlations almost perfectly match the overall correlations, but (for fun), we can also follow this up by checking how well the imputed and measured cell line response correlates in the cell lines that do overlap, that do appear in both datasets. *This will not be included in the paper draft*

```{r overlap analysis: imputed}
#use inner_join to find the CCLs that overlap with to GDSC, ie this time keeping ccls in GDSC that are also in CTRP
overlapping_CCLs <- inner_join(GDSCtomerge, Ccls_CTRP, by = "Harmonized Identifier")

#graph measured and Imputed response in all GSDC cell lines that overlap between GDSC and CTRP to GDSC. Probably should have made a graphing function, but oh well. 
Merge_preds_GDSC.IC50 %>% 
  semi_join(., overlapping_CCLs) %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity \n in Cell Lines That Overlap to CTRP", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) values from GDSC")

#graph measured and Imputed response in only BRCA GSDC cell lines that are unique to GDSC, n = 33
Merge_preds_GDSC.IC50 %>% 
  semi_join(., overlapping_CCLs) %>% 
  filter(`Cancer Statistics 2017 Sub-Category.x` == "Breast") %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity \n in Breast Cancer Cell Lines That Overlap Between GDSC and CTRP", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) values from GDSC")


```

Not surprising, these graphs look essentially the same as the overall and unique GDSC/ISS graphs. The correlation is no different or perhaps marginally better in the imputed and measured cell lines. Again, it makes sense that the correlation doesn't change whether the cell lines are unique or overlap in the two datasets since these two datasets are completely independent. 


However, it would be nice to have a standard to compare these results to, and we can look for one in the actual measured AUC and IC50 drug efficacy from CTRP and GDSC. That is, we can take the cell lines that overlap between GDSC/CTRP and compare their actual measured responses for each dataset. 


```{r overlap analysis: measured}

fig4b <- overlapping_CCLs %>% 
  ggplot(aes(x = Avg_AUC, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Measured CCL Sensitivity to AZD-1775 from GDSC and CTRP", 
         x = "Measured AUC Values from CTRP", 
         y = "Measured ln(IC50) values from GDSC") + 
    theme_bw(base_size = 14.6)

fig4b
fig4b <- fig4b + labs(title = NULL)


sup_4d <- overlapping_CCLs %>% 
  filter(`Cancer Statistics 2017 Sub-Category.x` == "Breast") %>% 
  ggplot(aes(x = Avg_AUC, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Measured BRCA CCL Sensitivity to AZD-1775 from GDSC and CTRP", 
         x = "Measured AUC Values from CTRP", 
         y = "Measured ln(IC50) values from GDSC")

sup_4d
sup_4d <- sup_4d + labs(title = NULL)


```


It is good to notice that the x-axis does change drastically in the "measured" values. It is not surprising since pRROphetic uses a linear ridge regression model that does tend to decrease the overall variability between the input value and the output value. If we remove some of the outliers, we can see the graph looks a bit better, but the correlation doesn't really improve or change (which makes sense since it is Spearman). We could also try to graph the relationships between the measured vs imputed on the same axis scale as the measured vs measured, but it looks very weird, seen below. 

```{r axis fixing}

#perhaps you could tell but there are some extreme values in the Measured AUC values that might be changing the data, so first I'll clip those values and see what happens
overlapping_CCLs %>% 
  filter(Avg_AUC < 15,
         Avg_AUC > 5) %>% 
  ggplot(aes(x = Avg_AUC, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Measured CCL Sensitivity to AZD-1775 from GDSC and CTRP", 
         x = "Measured AUC Values from CTRP", 
         y = "Measured ln(IC50) values from GDSC")


overlapping_CCLs %>% 
   ggplot(aes(x = Avg_AUC, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor(method = "spearman") + 
    labs(title = "Correlation of Measured CCL Sensitivity to AZD-1775 from GDSC and CTRP", 
         x = "Measured AUC Values from CTRP",
         y = "Measured ln(IC50) from GDSC") + 
     expand_limits(x = c(min(overlapping_CCLs$Avg_AUC),max(overlapping_CCLs$Avg_AUC)), 
                 y = c(min(GDSCtomerge$LN_IC50), max(GDSCtomerge$LN_IC50)))

#for comparison in the same chunk I'm graphing the measured and imputed response, but to do so on the same axis scale 

Merge_preds_GDSC.IC50 %>% 
  semi_join(., non_modeled_CCLs) %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor() + 
    labs(title = "Correlation of Imputed and Measured \n AZD-1775 Sensitivity in CCLs Unique to GDSC", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) from GDSC") + 
   expand_limits(x = c(min(overlapping_CCLs$Avg_AUC),max(overlapping_CCLs$Avg_AUC)), 
                 y = c(min(GDSCtomerge$LN_IC50), max(GDSCtomerge$LN_IC50)))

Merge_preds_GDSC.IC50 %>% 
  drop_na() %>% 
  ggplot(aes(x = `MK-1775`, y = LN_IC50)) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    ggpubr::stat_cor() + 
    labs(title = "Correlation of Imputed and Measured AZD-1775 Sensitivity in GDSC", 
         x = "Imputed Sensitivty Score", 
         y = "Measured ln(IC50) from GDSC") + 
     expand_limits(x = c(min(overlapping_CCLs$Avg_AUC),max(overlapping_CCLs$Avg_AUC)), 
                   y = c(min(GDSCtomerge$LN_IC50), max(GDSCtomerge$LN_IC50)))
``` 

```{r}

ggsave(plot = figure4a, filename = "./Figures/Figure4a.tiff", device = "tiff", width = 6, height = 4)
ggsave(plot = fig4b, filename = "./Figures/Figure4b.tiff", device = "tiff", width = 6, height = 4)

ggsave(plot = sfig4a, filename = "./Figures/Supplmental_figure4a.png", width = 6, height = 4)
ggsave(plot = sfig4b, filename = "./Figures/Supplmental_figure4b.png", width = 6, height = 4)

ggsave(plot = sup_fig4c, filename = "./Figures/Supplmental_figure4c.png", width = 6, height = 4)
ggsave(plot = sup_4d, filename = "./Figures/Supplmental_figure4d.png", width = 6, height = 4)




```




```{r}
print(sessionInfo())
```
