==> 04_Analyze_TCGA-Imputation-Data.Rmd <==
---
title: "Analyzing TCGA (patient) Imputation Data"
output: html_notebook
author:
  - "Robert F. Gruener"
date: "date last edited: `r format(Sys.time(), '%d %B, %Y')`"
---
```{r, include=FALSE}
knitr::opts_chunk$set(warning = F, message = FALSE, cache = T)
```

Analysis outline: 

  * Install TCGAbiolinks to get TCGA clinical data
  * Format BRCA predictions, merge with clinical data
  * Perform t-tests and correct for multiple testing 
  * Graph results for AZD-1775
  


```{r install packages, include= FALSE}
library(TCGAbiolinks)
library(broom)
library(readxl)
library(grid)
library(ggrepel)
library(tidyverse)

theme_set(theme_bw())
```



```{r download TCGA clinical data}
query <- GDCquery(project = "TCGA-BRCA", 
                  data.category = "Clinical",
                  data.type = "Clinical Supplement", 
                  data.format = "BCR Biotab")
GDCdownload(query)
clinical.BCRtab.all <- GDCprepare(query)
names(clinical.BCRtab.all)
BRCA_clinical <- clinical.BCRtab.all$clinical_patient_brca %>%
    dplyr::select(bcr_patient_barcode, er_status_by_ihc, er_status_ihc_Percent_Positive, pr_status_by_ihc, 
                  pr_status_ihc_percent_positive, her2_status_by_ihc, her2_fish_status, ) %>%
    dplyr::slice(-1:-2) %>% 
    type_convert() %>% 
    mutate(HER2_Subtype = if_else(condition = her2_fish_status %in% c("Positive", "Negative"), her2_fish_status, her2_status_by_ihc)) %>% 
    filter(er_status_by_ihc %in% c("Positive", "Negative"), 
           pr_status_by_ihc %in% c("Positive", "Negative"), 
           HER2_Subtype %in% c("Positive", "Negative")) %>% 
    mutate(Clin_Subtype = if_else(condition = er_status_by_ihc == "Positive" | pr_status_by_ihc == "Positive", 
                                  true = if_else(condition = HER2_Subtype == "Positive", true = "ER/PR+, HER2+", false = "ER/PR+"),
                                  false = if_else(condition = HER2_Subtype == "Positive", true = "HER2+", false = "TNBC"))) %>% 
    mutate(TNBC_status = if_else(Clin_Subtype == "TNBC", "TNBC", "RPBC"))

TCGA_subs <- TCGAquery_subtype(tumor = "brca") #easily get TCGA PAM50 status this way

BRCA_pam50 <- TCGA_subs %>% dplyr::select(patient, PAM50 = BRCA_Subtype_PAM50) %>% filter(PAM50 != "NA")

BRCA_clinical <- full_join(BRCA_clinical, BRCA_pam50, by = c("bcr_patient_barcode" = "patient")) 

```

Ki-67 staining is not available in TCGA, so we cannot do the traditional Lum A and Lum B groups for the clinical subtyping of ER+ tumors. However, this doesn't impact us too much as there we are focused on TNBC which can be easily defined in our dataset.  




### Filtering drugs based on the cross-validation results
It shouldn't be expected that every drug should be modeled accurately whether this is due to inconsistency of the screening data or problems with sample size or other model variables. 

```{r}
CTRP_cv <- read_csv("./DataOut/20FoldCV_CTRP.csv")

CTRP_cv <- CTRP_cv %>% 
  filter(!grepl(pattern = "mol/mol", Drug))

cv_cutoff <- CTRP_cv %>% 
  filter(Spearman_Pvalue < 0.05,
         Spearman_Correlation > 0)

bad_cv_drugs <- anti_join(CTRP_cv, cv_cutoff)
bad_cv_drugs <- bad_cv_drugs$Drug


sup1 <- CTRP_cv %>% ggplot(aes(x = Spearman_Correlation)) + 
  geom_histogram() + 
  geom_vline(xintercept = min(cv_cutoff$Spearman_Correlation), color = "red") + 
  labs(x = "Spearman R from CTRP cross-validation", y = "Number of Drugs") +
  theme_bw(base_size = 18)
sup1
ggsave(plot = sup1, device = "eps", filename = "supplemental_figure_1.eps")

```

```{r}
drug_targets <- read_csv("./Included_Data/Drug_targets(review).csv") # From Ling and Gruener et al

median(cv_cutoff$Spearman_Correlation)

cv_targets <- cv_cutoff %>% 
  left_join(., drug_targets, by = c("Drug" = "Name in Database")) %>% 
  dplyr::select(-Database) %>% 
  distinct() %>%  
  filter(!Targets %in% c("#N/A", "0"))

length(unique(unlist(strsplit(cv_targets$Targets, split = "\\|"))))


```

### Merging Predictions with clinical data
Next we need to load and format the prediction dataset, then we merge it with the clinical data we previously downloaded

```{r format predictions, include = FALSE}
load("./DataOut/2015CTRP-Models_on_TCGA_BRCA.RData")
BRCA_preds <- as_tibble(t(CTRPDrugPredictions_mat), rownames = "Patient_ID") 
BRCA_preds <- BRCA_preds %>% 
  filter(grepl(Patient_ID, pattern = "\\.01A\\.|\\.01B\\.")) %>% #remove normal or metastatic samples
  mutate(Patient_ID = gsub(x = strtrim(Patient_ID, width = 12), pattern = "\\.", replacement = "-")) %>% 
  gather(key = "Drug", value = "PSS", -Patient_ID) %>% #turns the table into a long format
  mutate(Drug = replace(Drug, Drug == "MK-1775", "AZD-1775")) %>%  #replaces MK-1775 as AZD-1775 since AZD-1775 is the more common name currently even though this is not how it is written in the database
  filter(!Drug %in% bad_cv_drugs,
         !grepl(pattern = "mol/mol", Drug)) #filter out drugs with poor CV results


#join with clinical information
preds.subtype <- left_join(BRCA_preds, BRCA_clinical, by = c("Patient_ID" = "bcr_patient_barcode")) %>% 
  drop_na(TNBC_status, PAM50)

write_csv(preds.subtype, file = "Supplemental_table1.csv")

```



### Tamoxifen imputated sensitivity and ER+ breast cancers
We have reported on HER2 inhibitors in HER2 positive cancers previously, we can look for another proof-of-concept by looking at which drugs associate with the ER-positive BRCA cohort. We'll do a similar test, stratify between ER+ and non-ER+ and perform a t-test between the two, and look for the compounds that associate with the ER+ subset. 


```{r patient information}
#first how many patients 
preds.subtype %>% select(Patient_ID, Clin_Subtype) %>% distinct() %>% count(Clin_Subtype)
preds.subtype %>% select(Patient_ID, PAM50) %>% distinct() %>% count(PAM50)

preds.subtype %>% select(Drug) %>% distinct() %>% count(Drug)



```

```{r ER Drugs}
###########ER and Tamoxifen Response#################

#ER and Tamoxifen stats:

ER_t.tst_preds <- preds.subtype %>% 
  group_by(Drug) %>% 
  drop_na(er_status_by_ihc) %>% 
  do(tidy(t.test(PSS~er_status_by_ihc, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value)

dplyr::filter(ER_t.tst_preds, estimate > 0)
#ER positivity associates with Imputed Tamoxifen Response, it is the 3th most significant result
```


```{r}

lum_t.tst_preds <- preds.subtype %>%
  filter(Clin_Subtype %in% c("ER/PR+", "HER2+")) %>% 
  group_by(Drug) %>% 
  do(tidy(t.test(PSS~Clin_Subtype, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value) 

lum_t.tst_preds %>% 
  filter(estimate < 0) %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:3) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = lum_t.tst_preds) +
    geom_point(color = "red", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in RPBC                                                                                                            Drugs More Effective in TNBC") + 
    theme(plot.title = element_text(hjust = 0.5))

lum_t.tst_preds <- preds.subtype %>%
  filter(Clin_Subtype %in% c("ER/PR+", "TNBC")) %>% 
  group_by(Drug) %>% 
  do(tidy(t.test(PSS~Clin_Subtype, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value)

lum_t.tst_preds %>% 
  filter(estimate < 0) %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:3) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = lum_t.tst_preds) +
    geom_point(color = "red", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in RPBC                                                                                                            Drugs More Effective in TNBC") + 
    theme(plot.title = element_text(hjust = 0.5))


lum_t.tst_preds <- preds.subtype %>%
  filter(Clin_Subtype %in% c("ER/PR+", "HER2+", "TNBC")) %>% 
  mutate(luminal = grepl("ER/PR+", x = .$Clin_Subtype)) %>% 
  group_by(Drug) %>% 
  do(tidy(t.test(PSS~luminal, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value)

lum_t.tst_preds %>% 
  filter(estimate < 0) %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:4) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = lum_t.tst_preds) +
    geom_point(color = "red", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in RPBC                                                                                                            Drugs More Effective in TNBC") + 
    theme(plot.title = element_text(hjust = 0.5))



lum_t.tst_preds <- preds.subtype %>%
  filter(PAM50 %in% c("LumA", "Basal")) %>% 
  group_by(Drug) %>% 
  do(tidy(t.test(PSS~PAM50, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value)

lum_t.tst_preds %>% 
  filter(estimate > 0) %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:3) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = lum_t.tst_preds) +
    geom_point(color = "red", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in Basal                                                                                                            Drugs More Effective in Luminal") + 
    theme(plot.title = element_text(hjust = 0.5))

```

```{r}
lum_t.tst_preds <- preds.subtype %>%
  filter(Clin_Subtype %in% c("ER/PR+", "TNBC")) %>% 
  mutate(Clin_Subtype = factor(Clin_Subtype, levels = c("TNBC", "ER/PR+"))) %>% 
  group_by(Drug) %>% 
  do(tidy(t.test(PSS~Clin_Subtype, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(bonf_p.value) 

e_10_cutoff <- sort(lum_t.tst_preds$estimate[lum_t.tst_preds$estimate > 0], 
                    decreasing = T)[round(sum(lum_t.tst_preds$estimate > 0)*0.1, digits = 0)] 
p_10_cutoff <- sort(lum_t.tst_preds$bonf_p.value[lum_t.tst_preds$estimate > 0], 
                    decreasing = F)[round(sum(lum_t.tst_preds$estimate > 0)*0.1, digits = 0)]

top_lum_drugs <- lum_t.tst_preds %>% 
  filter(estimate > 0, 
         bonf_p.value < p_10_cutoff)

lum_t.tst_preds %>% 
  filter(estimate > 0) %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:11) %>% 
  ggplot(mapping = aes(x = 1, y = log10_p.correct, label = Drug)) + 
    geom_point(data = filter(lum_t.tst_preds, estimate > 0)) +
    geom_point(color = "blue", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in RPBC                                                                                                            Drugs More Effective in TNBC") + 
    theme(plot.title = element_text(hjust = 0.5))


lum_t.tst_preds %>% 
  filter(estimate > 0) %>% 
  mutate(label = if_else(bonf_p.value < p_10_cutoff, Drug, ""),
         Signific = if_else(bonf_p.value < p_10_cutoff, "Lead", "")) %>% 
  ggplot(mapping = aes(x = 1, y = log10_p.correct, label = label)) + 
  geom_point(aes(color = Signific)) + 
  scale_color_manual(values = c("black", "blue")) + 
  geom_text_repel(nudge_x = 0.01, force = 3) + 
  scale_x_continuous(name = "Drugs with Predicted Preferential \nAcitivity in HR+ Breast Cancer", 
                     limits = c(1, 1.08), breaks = NULL) + 
    labs(y = "-log10(adjusted p-value)") +
   theme_classic() +
    theme(axis.title.x = element_text(hjust = 0),
          legend.position = "none") 
    
  
  

```


```{r HER2 Drugs}
########### HER2 Drugse#################


HER2_t.tst_preds <- preds.subtype %>% 
  group_by(Drug) %>% 
  drop_na(er_status_by_ihc) %>% 
  do(tidy(t.test(PSS~HER2_Subtype, data = .))) %>% 
  ungroup() %>% 
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>%
  mutate(log10_p.correct = -log10(bonf_p.value)) %>% 
  arrange(log10_p.correct) %>% 
  filter(estimate > 0)

head(HER2_t.tst_preds)

#ER positivity associates with Imputed Tamoxifen Response, it is the 5th most significant result
```


### Identifying Compounds Predicted To Be More Effective in TNBC


To look for compounds that are more effective in TNBC vs other breast cancers (i.e. receptor positive or RPBC), we can subset the patients into their corresponding TNBC status and perform t-tests on the imputed drug data for each drug. Then we can correct for multiple test corrections. 

```{r Statistical testing}
t.tst_preds.subtype <- preds.subtype %>% 
  group_by(Drug) %>%  #group by drug so that each statistical test is performed for each drug
  do(tidy(t.test(PSS~TNBC_status, data = .))) %>% #way to perform t-tests and get the results from the t-test into a tbl format
  ungroup() %>% #need to ungroup so that when we adjust we are adjusting everything, otherwise the group is set to make each drug independent
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) 

#We could also look at the results via the PAM50 subgroups. I'm keeping this analysis here but some of this may become supplemental.  
PAM50_preds.subtype <- preds.subtype %>%
  mutate(Basal_status = if_else(PAM50 == "Basal", "Basal", "non-Basal")) %>% 
  group_by(Drug) %>%
  do(tidy(t.test(PSS~Basal_status, data = .))) %>% #way to perform t-tests and get the results from the t-test into a tbl format
  ungroup() %>% #need to ungroup so that when we adjust we are adjusting everything, otherwise the group is set to make each drug independent
  mutate(bonf_p.value = p.adjust(p.value, method = "bonferroni")) %>% 
  mutate(log10_p.correct = -log10(bonf_p.value)) 


```




```{r, echo= FALSE}
#total number of significant results
print("total")
sum(t.tst_preds.subtype$bonf_p.value < 0.01)

#total number for TNBC
print("significant results in TNBC")
sum(t.tst_preds.subtype[t.tst_preds.subtype$estimate > 0,]$bonf_p.value < 0.01)


t.tst_preds.subtype %>% 
  ggplot() +
    geom_histogram(mapping = aes(x = log10_p.correct)) +
    geom_vline(xintercept = -log10(0.01), color = "red") +
    labs(title = "Histogram of T-test Results by Signifcance", 
         y = "Number of Drugs", 
         x = "-log10(bonferroni-adjusted p-value)")

```

As a note: 
Based on t-tests, even with a bonferroni correction, most of the values are statistically signifcant. From this table `r sum(t.tst_preds.subtype$bonf_p.value < 0.01/length(unique(t.tst_preds.subtype$Drug)), na.rm = TRUE)`% are statistically significant at a corrected p-value of 0.01 out of the `r length(unique(t.tst_preds.subtype$Drug))`total drugs. This does indicte that we are getting an enrichment of likely false positive results, so the best thing to do would be to just consider the most significant results, essentially placing a much higher significant cutoff threshold. This should be investigated in future analysis. 



To visualize the t-test results, we can create a volcano plot and look for the most significant results: 

```{r volcano plot of Estimate vs p-value}

t.tst_preds.subtype %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:3) %>% 
  bind_rows(slice(arrange(filter(t.tst_preds.subtype, estimate < 0),
                                    desc(log10_p.correct)),
                            1:2)) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = t.tst_preds.subtype) +
    geom_point(color = "red", size = 2.2) +
    geom_point(data = slice(arrange(filter(t.tst_preds.subtype, estimate < 0),
                                    desc(log10_p.correct)),
                            1:2), 
               color = "blue", size = 2.2, mapping = aes(label = Drug)) +
    geom_text_repel(nudge_y = 3) + 
    labs(title = "Volcano Plot of TNBC vs RPBC T-Test Results \n P-values and Estimate Scores for All Drugs", 
         y = "-log10(p.value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = "Drugs More Effective in RPBC                                                                                                            Drugs More Effective in TNBC") + 
    theme(plot.title = element_text(hjust = 0.5))

#making version of plot for text
dir.create(path = "./Figures")

fig2a <- t.tst_preds.subtype %>% 
  arrange(desc(log10_p.correct)) %>% 
  dplyr::slice(1:3) %>% 
  ggplot(mapping = aes(x = estimate, y = log10_p.correct, label = Drug)) + 
    geom_point(data = t.tst_preds.subtype) +
    geom_point(color = "red", size = 2.2) + 
    geom_text_repel(nudge_y = 3) + 
    theme_bw(base_size = 16) +
    labs(y = "-log10(adjusted p-value)", 
         x = "Mean Difference (Imputed Effect RPBC - TNBC)", 
         caption = c("Drugs More\nEffective in TNBC", " Drugs More\nEffective in RPBC")) + 
    theme(plot.caption = element_text(hjust=c(1, -0.1))) 
    
fig2a

ggsave(filename = "./Figures/Figure2a.TIFF", device = "tiff", plot = fig2a, height = 5.5, width = 7)
    
```

Highlighted in red are the 3-most significant results. Due to toxicity reasons, we chose not to look at AZD7762 or leptomycin B. AZD-1775 has some evidence that suggests its use in TNBC, although not robust, and it is the third most significant result and has a good toxicity profile already. Additionally, our first analysis looked at ANOVA results in the PAM50 subtype and AZD-1775 definitely associated with the Basal PAM50 subtype (Basal is essentially TNBC, but there are some differences especially with the LAR TNBC subtype, but overall basal roughly corresponds with TNBC). But we can look at the top drugs that associate with TNBC by effect size and p-value. 

## Top TNBC Associated Drugs 

We want to look drugs that are imputed to be more effective, more selective for TNBC by both effect size and p-value cutoffs. As mentioned earlier, we do see many drugs that are significant at the p < 0.05 cut-off. Thus we will look at effect size as well and use compounds in the top 10% in terms of effect size and significance.

We will use the review co-authored by Alex Ling and myself (Robert F. Gruener) to obtain drug target and pathway information on. 

```{r drug targets}

targets_t.test <-  t.tst_preds.subtype %>%
  mutate(Drug = replace(Drug, Drug == "AZD-1775", "MK-1775")) %>% #replacng MK1775 for AZD1775 since it will be easier to match using "Name in Database" instead of all names
  dplyr::select(Drug, estimate, bonf_p.value, log10_p.correct) %>% 
  left_join(., drug_targets, by = c("Drug" = "Name in Database")) %>% 
  dplyr::select(-Database) %>% 
  distinct() %>% 
  mutate(Drug = replace(Drug, Drug == "MK-1775", "AZD-1775"), 
         Drug = replace(Drug, Drug == "triazolothiadiazine", "ML030")) %>% 
  arrange(bonf_p.value)

targets_t.test

```

```{r top drugs MOA}

#For the table in the text of the paper, I'm going to include only the top 10% based on both effect size and significance, sorted by mechanism of action. Here is roughly what I'll be looking for, some edits in excel may be called for:

#Here I'm more concerned about the top compounds based on effect size and p-value, so find the cutoff based on the ~290 sig drugs
e_10_cutoff <- sort(t.tst_preds.subtype$estimate, 
                    decreasing = T)[round(sum(t.tst_preds.subtype$estimate > 0)*0.1, digits = 0)] 
p_10_cutoff <- sort(abs(t.tst_preds.subtype$bonf_p.value), 
                    decreasing = F)[round(sum(t.tst_preds.subtype$estimate > 0)*0.1, digits = 0)]

temp_top_targets_t.tst <- targets_t.test %>% 
 filter(estimate > e_10_cutoff,
         bonf_p.value < p_10_cutoff) 

nrow(temp_top_targets_t.tst) #there are 17 drugs that fit these cut-offs

# Second, some compounds may not have MOA information, but most of these do have gene targets listed. I will use the gene target information and do some quick validation (read: googling) on the compound and targets to be able to convert them into a 
missing_MOA <- temp_top_targets_t.tst[temp_top_targets_t.tst$MOA %in% c("0","#N/A"),]
missing_MOA
found_MOA <- vector(mode= "character", length = nrow(missing_MOA))
found_MOA <- c("translation (eIF4F complex) inhibiitor", "CDK inhibitor", "phosphodiesterase inhibitor", "translation (eIF4F complex) inhibiitor")
names(found_MOA) = c("CR-1-31B", "BRD-K30748066", "ML030", "SR-II-138A")

#update the original df with this information too
targets_t.test$MOA[targets_t.test$Drug %in% names(found_MOA)] <- found_MOA

```


```{r Table 1: Lead Candidates for TNBC by MOA}
#Finally, I ideally want this as simple as possible but some have multiple MOA listed. I will use the first one listed as the primary one, since it will be easiest thi s way. However, this isn't necessarily the best for this compound. It appears when there are multiple, it is listed in alphabetical order; however, this doesn't appear to matter for the MOAs that will be included in table 1. The exception is that a PLK inhibitor, since "cell cycle inhibitor" is listed first, this will be changed in the next line

targets_t.test$MOA[targets_t.test$MOA == "cell cycle inhibitor|PLK inhibitor"] <- "PLK inhibitor|cell cycle inhibitor"
targets_t.test$MOA <- gsub(pattern = "\\|.*$", replacement = "", x = targets_t.test$MOA) #only keeps the first MOA listed
targets_t.test <- targets_t.test %>% count(MOA, name = "Total_MOA_Count") %>% left_join(targets_t.test, .)

#now that there have been updates to the original df, rewrite the top_targets_t.tst
top_targets_t.tst <- targets_t.test %>% 
 filter(estimate > e_10_cutoff,
         bonf_p.value < p_10_cutoff)

#now for the top compounds get a table that is MOA focused with a "total MOA" count and a count for the number of top compounds 

top_TNBC_drugs <- top_targets_t.tst %>% 
  count(MOA, Total_MOA_Count, sort = T) %>% 
  left_join(., y = top_targets_t.tst, by = c("MOA", "Total_MOA_Count")) %>% 
  dplyr::group_by(MOA) %>% 
  mutate(drugs = paste(Drug, collapse = ", ")) %>% 
  ungroup() %>% 
  filter(estimate >= e_10_cutoff,
        bonf_p.value <= p_10_cutoff) %>%
  arrange(bonf_p.value) %>% 
  select(MOA, Total_MOA_Count, n, drugs) %>% 
  distinct()

top_TNBC_drugs

#save the top_compounds as a table for the paper
write_csv(top_TNBC_drugs, path = "./DataOut/Table1_draft.csv")

#lets save the top_targets_t.test as a supplementary table
targets_t.test %>% 
  mutate(MOA = na_if(MOA, MOA %in% c("0", "#N/A"))) %>% 
  arrange(bonf_p.value) %>% 
  dplyr::select(Drug, "Pubchem ID" = `Harmonized ID (Pubchem ID)`, "Effect Size" = estimate, 
                "Bonferroni Adjusted p-value" = bonf_p.value, MOA, Targets, 
                `Clinical Phase`, `Source for Clinical Information`, "N Drugs with same MOA" = Total_MOA_Count) %>% 
  write_csv(., path = "./DataOut/Supplementary Table 1.csv")

```


### Graphing results in BRCA subtypes 

We can look how certain drugs are 

```{r graphing function 1}

library(ggpubr) #ggpubr has some nice default settings for adding comparisons
library(ggplot2)

Sensitivity_grapher <- function(drug, comparison, group = "TNBC"){

temp.preds.subtype <- preds.subtype %>% 
    filter(Drug == drug,
           PAM50 != "Normal") 
  
subtypes <- unique(eval(parse(text = paste("temp.preds.subtype$", comparison)))) #removing PAM50 normal
test <- combn(subtypes, 2) #get all the comparisons
test_list <- split(test, rep(1:ncol(test), each = 2))
comp_list <- test_list[grepl(pattern = group , test_list)] #get all the comparisons between the group of interest (TNBC)

preds.subtype <- preds.subtype %>% 
    filter(Drug == drug) %>% 
    filter(!!rlang::sym(comparison) %in% subtypes) %>% 
    mutate(PAM50 = factor(PAM50, levels = c("LumA", "LumB", "Her2", "Basal")), 
           Clin_Subtype = factor(Clin_Subtype, levels = c("ER/PR+", "ER/PR+, HER2+", "HER2+", "TNBC")))
          

box_labs <- preds.subtype %>% #placing the number of patients in each subtype into the boxes
    count_(comparison) %>% 
    arrange_(comparison) %>% 
    .$n

print(                       #graphing function that graphs boxplots with comparisons 
  preds.subtype %>% 
  ggboxplot(x = comparison, y = "PSS", fill = comparison) +
      theme_bw(base_size = 16) + 
      labs(title = paste(drug, "'s Imputed Sensitivity in TCGA BRCA Cohort by Receptor Status", sep = ""),
           y = "Imputed Sensitivity Score", 
           x = "BRCA Subtype") +
      theme(legend.position = "none",
            plot.title = element_text(hjust = 0.5))+
    stat_compare_means(comparisons = comp_list) + 
    grids() + 
    stat_summary(geom = "text", label = paste("n =", box_labs), vjust = -0.5)
      
)
  
}

#now use this graphing function to graph AZD-1775's response in different BRCA subtypes
azd_clin <- Sensitivity_grapher(drug = "AZD-1775", comparison = "Clin_Subtype")
azd_pam <- Sensitivity_grapher(drug = "AZD-1775", comparison = "PAM50", group = "Basal") + 
  labs(title = "AZD-1775's Imputed Sensitivity in TCGA BRCA Cohort by PAM50 Status", 
       x = "PAM50 Subtype") #need to change the titles and axis when I change groups, may need to change that in the future

#can also look at lapatinib and tamoxifen 
Sensitivity_grapher(drug = "leptomycin B", comparison = "Clin_Subtype", group = "TNBC")


tam_clin <- Sensitivity_grapher(drug = "tamoxifen", comparison = "Clin_Subtype", group = "ER/PR+") 

tam_pam <- Sensitivity_grapher(drug = "tamoxifen", comparison = "PAM50", group = "Lum")

```

```{r AZD graphs}
azd_clin <- azd_clin + 
  labs(title = NULL, 
       y = "AZD-1775 Imputed Sensitivity Score",
       x = "TCGA Patients by Clinical BRCA Subtype") 
azd_clin

ggsave(filename = "./Figures/Figure2b.tiff", device = "tiff", plot = azd_clin)


azd_pam <- azd_pam + 
  labs(title = NULL, 
       y = "AZD-1775 Imputed Sensitivity Score",
       x = "TCGA Patients by PAM50 Subtype")
azd_pam

ggsave(filename = "./Figures/Figure2c.tiff", device = "tiff", plot = azd_pam)
```

```{r Tamoxifen Graphs}

tam_clin <- tam_clin +
    labs(title = "Tamoxifen's Imputed Sensitivity in TCGA BRCA Cohort by Receptor Status")
tam_clin
ggsave(filename = "./Figures/Supplement_2a.png", plot = tam_clin)

tam_pam <- tam_pam + 
  labs(title = "Tamoxifen's Imputed Sensitivity in TCGA BRCA Cohort by PAM50 Subtype")
tam_pam
ggsave(filename = "./Figures/Supplement_2b.png", plot = tam_pam)

```



```{r}
print(sessionInfo())
```
